Allow OpenTSDB URL to be set at scrape time
Instead of setting the OpenTSDB URL as a command-line argument, allow it
to be set using the `target` parameter at scrape time.

This change makes it possible to specify OpenTSDB targets using
Prometheus' service discovery in combination with relabelling, which is
important when you have multiple tsd instances.

Also removes the `web.telemetry-path` command-line argument in favour of
using `/metrics` for metrics about this exporter and `/opentsdb` for
metrics pulled from the OpenTSDB stats API.

This pattern is borrowed from the Blackbox and SNMP exporters, e.g.:
https://github.com/prometheus/snmp_exporter/blob/aa8c078383e41412ab346e8053b88a9cae905ca8/main.go