Don't output error in HTTP response body
The exporter allows arbitrary URLs as targets, so we should avoid
returning detailed error information in the response body in case an
attacker were to use the exporter to attempt to discover/query HTTP
servers in networks with poor isolation between services.

Instead, the user should check the exporters logs, which do include the
detailed error message.